requirements.txt
Werkzeug==0.16.0
cryptography
requests
Flask
Flask-HTTPAuth
flask-restplus
flask-jwt-extended
python-consul
pyaml
bcrypt
pytest